configure routes, add debug page
