+++ packages/react-app/src/views/Debug.jsx
@@ -1,0 +1,221 @@
+import React, { useCallback, useEffect, useState } from "react";
+import { getDefaultProvider, InfuraProvider, JsonRpcProvider, Web3Provider } from "@ethersproject/providers";
+import { Row, Col, Button } from "antd";
+import Web3Modal from "web3modal";
+import WalletConnectProvider from "@walletconnect/web3-provider";
+import { useUserAddress } from "eth-hooks";
+import { useExchangePrice, useGasPrice, useUserProvider, useContractLoader, useContractReader, useBalance, useEventListener } from "../hooks";
+import { Header, Account, Faucet, Ramp, Contract, GasGauge } from "../components";
+import { Transactor } from "../helpers";
+import Hints from "../Hints";
+/*
+    Welcome to 🏗 scaffold-eth !
+
+    Code:
+    https://github.com/austintgriffith/scaffold-eth
+
+    Support:
+    https://t.me/joinchat/KByvmRe5wkR-8F_zz6AjpA
+    or DM @austingriffith on twitter or telegram
+
+    You should get your own Infura.io ID and put it in `constants.js`
+    (this is your connection to the main Ethereum network for ENS etc.)
+*/
+import { INFURA_ID, ETHERSCAN_KEY } from "../constants";
+
+// 🛰 providers
+console.log("📡 Connecting to Mainnet Ethereum");
+const mainnetProvider = getDefaultProvider("mainnet", { infura: INFURA_ID, etherscan: ETHERSCAN_KEY, quorum: 1 });
+// const mainnetProvider = new InfuraProvider("mainnet",INFURA_ID);
+// const mainnetProvider = new JsonRpcProvider("https://mainnet.infura.io/v3/5ce0898319eb4f5c9d4c982c8f78392a")
+// ( ⚠️ Getting "failed to meet quorum" errors? Check your INFURA_ID)
+
+// 🏠 Your local provider is usually pointed at your local blockchain
+// as you deploy to other networks you can set REACT_APP_PROVIDER=https://dai.poa.network in packages/react-app/.env
+const localProviderUrl = process.env.REACT_APP_PROVIDER ? process.env.REACT_APP_PROVIDER : "http://localhost:8545"; // https://dai.poa.network
+console.log("🏠 Connecting to provider:", localProviderUrl);
+const localProvider = new JsonRpcProvider(localProviderUrl);
+
+/*
+  Web3 modal helps us "connect" external wallets:
+*/
+const web3Modal = new Web3Modal({
+  // network: "mainnet", // optional
+  cacheProvider: true, // optional
+  providerOptions: {
+    walletconnect: {
+      package: WalletConnectProvider, // required
+      options: {
+        infuraId: INFURA_ID,
+      },
+    },
+  },
+});
+
+const logoutOfWeb3Modal = async () => {
+  await web3Modal.clearCachedProvider();
+  setTimeout(() => {
+    window.location.reload();
+  }, 1);
+};
+
+function Debug() {
+  const [injectedProvider, setInjectedProvider] = useState();
+  /* 💵 this hook will get the price of ETH from 🦄 Uniswap: */
+  const price = useExchangePrice(mainnetProvider); //1 for xdai
+
+  /* 🔥 this hook will get the price of Gas from ⛽️ EtherGasStation */
+  const gasPrice = useGasPrice("fast"); //1000000000 for xdai
+
+  // For more hooks, check out 🔗eth-hooks at: https://www.npmjs.com/package/eth-hooks
+
+  // Use your injected provider from 🦊 Metamask or if you don't have it then instantly generate a 🔥 burner wallet.
+  const userProvider = useUserProvider(injectedProvider, localProvider);
+  const address = useUserAddress(userProvider);
+
+  const transactor = Transactor(userProvider)
+
+  // 🏗 scaffold-eth is full of handy hooks like this one to get your balance:
+  const yourLocalBalance = useBalance(localProvider, address);
+  console.log("💵 yourLocalBalance",yourLocalBalance)
+
+  // just plug in different 🛰 providers to get your balance on different chains:
+  const yourMainnetBalance = useBalance(mainnetProvider, address);
+  console.log("💵 yourMainnetBalance",yourMainnetBalance)
+
+  // Load in your local 📝 contract and read a value from it:
+  const readContracts = useContractLoader(localProvider)
+  console.log("📝 readContracts",readContracts)
+
+  // keep track of a variable from the contract in the local React state:
+  const purpose = useContractReader(readContracts,"YourContract", "purpose")
+  console.log("🤗 purpose:",purpose)
+
+  // If you want to make 🔐 write transactions to your contracts, use the userProvider:
+  const writeContracts = useContractLoader(userProvider)
+  console.log("🔐 writeContracts",writeContracts)
+
+  //📟 Listen for broadcast events
+  //const setPurposeEvents = useEventListener(readContracts, "YourContract", "SetPurpose", localProvider, 1);
+  //console.log("📟 SetPurpose events:",setPurposeEvents)
+
+  const loadWeb3Modal = useCallback(async () => {
+    const provider = await web3Modal.connect();
+    setInjectedProvider(new Web3Provider(provider));
+  }, [setInjectedProvider]);
+
+  useEffect(() => {
+    if (web3Modal.cachedProvider) {
+      loadWeb3Modal();
+    }
+  }, [loadWeb3Modal]);
+
+  return (
+    <div className="App">
+      {/* ✏️ Edit the header and change the title to your project name */}
+      <Header />
+
+      {/*
+        ⚙️ Here is an example button that sets the purpose in your smart contract:
+
+
+      */}
+
+      <Button onClick={()=>{
+        transactor( writeContracts.CLR.startRound("1000") )
+      }}>START</Button>
+
+       {/* 👨‍💼 Your account is in the top right with a wallet at connect options */}
+      <div style={{ position: "fixed", textAlign: "right", right: 0, top: 0, padding: 10 }}>
+        <Account
+          address={address}
+          localProvider={localProvider}
+          userProvider={userProvider}
+          mainnetProvider={mainnetProvider}
+          price={price}
+          web3Modal={web3Modal}
+          loadWeb3Modal={loadWeb3Modal}
+          logoutOfWeb3Modal={logoutOfWeb3Modal}
+        />
+      </div>
+
+      {/*
+          🎛 this scaffolding is full of commonly used components
+          this <Contract/> component will automatically parse your ABI
+          and give you a form to interact with it locally
+      */}
+
+      <Contract name="CLR" signer={userProvider.getSigner()} provider={localProvider} address={address} />
+
+      {/*
+
+        📑 Maybe display a list of events?
+
+        <div style={{ width:600, margin: "auto" }}>
+        <List
+          bordered
+          dataSource={setPurposeEvents}
+          renderItem={item => (
+            <List.Item>
+              {item[0]} =>
+              {item[1]}
+            </List.Item>
+          )}
+        />
+      </div>
+
+      */}
+
+
+      {/* 🗑 Throw these away once you have 🏗 scaffold-eth figured out: */}
+      <Hints address={address} yourLocalBalance={yourLocalBalance} price={price} mainnetProvider={mainnetProvider} />
+
+
+      {/* 🗺 Extra UI like gas price, eth price, faucet, and support: */}
+       <div style={{ position: "fixed", textAlign: "left", left: 0, bottom: 20, padding: 10 }}>
+         <Row align="middle" gutter={[4, 4]}>
+           <Col span={8}>
+             <Ramp price={price} address={address} />
+           </Col>
+
+           <Col span={8} style={{ textAlign: "center", opacity: 0.8 }}>
+             <GasGauge gasPrice={gasPrice} />
+           </Col>
+           <Col span={8} style={{ textAlign: "center", opacity: 1 }}>
+             <Button
+               onClick={() => {
+                 window.open("https://t.me/joinchat/KByvmRe5wkR-8F_zz6AjpA");
+               }}
+               size="large"
+               shape="round"
+             >
+               <span style={{ marginRight: 8 }} role="img" aria-label="support">
+                 💬
+               </span>
+               Support
+             </Button>
+           </Col>
+         </Row>
+
+         <Row align="middle" gutter={[4, 4]}>
+           <Col span={24}>
+             {
+               /*  if the local provider has a signer, let's show the faucet:  */
+               localProvider && !process.env.REACT_APP_PROVIDER && price > 1 ? (
+                 <Faucet localProvider={localProvider} price={price} />
+               ) : (
+                 ""
+               )
+             }
+           </Col>
+         </Row>
+       </div>
+
+
+
+
+    </div>
+  );
+}
+
+export default Debug;
