+++ packages/react-app/src/views/index.js
@@ -1,0 +1,1 @@
+export { default as Debug } from "./Debug";
